Move DeleteDepartmentCommand to commands.department
