+++ src/test/java/seedu/sudohr/logic/commands/DeleteDepartmentCommandTest.java
@@ -11,6 +11,7 @@ import static seedu.sudohr.testutil.TypicalDepartments.getTypicalSudoHr;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.sudohr.logic.commands.department.DeleteDepartmentCommand;
 import seedu.sudohr.model.Model;
 import seedu.sudohr.model.ModelManager;
 import seedu.sudohr.model.UserPrefs;
