+++ src/test/java/seedu/sudohr/logic/parser/DeleteDepartmentCommandParserTest.java
@@ -12,7 +12,7 @@ import static seedu.sudohr.logic.parser.CommandParserTestUtil.assertParseSuccess
 
 import org.junit.jupiter.api.Test;
 
-import seedu.sudohr.logic.commands.DeleteDepartmentCommand;
+import seedu.sudohr.logic.commands.department.DeleteDepartmentCommand;
 import seedu.sudohr.model.department.DepartmentName;
